Remove listening app as it has been merged with the tutor app
The functionality of the listening app has been integrated into the tutor
app to streamline the codebase and reduce redundancy. All relevant models,
views, templates, and services have been migrated or refactored into
the tutor app. This change simplifies maintenance and improves the
overall structure of the project.

Note that we are keeping the migrations for the listening app, as they are
needed for database integrity until we migrate the production instance.